re-organize task notifications into package
