+++ tunga_tasks/management/commands/tunga_manage_task_progress.py
@@ -4,7 +4,7 @@ from dateutil.relativedelta import relativedelta
 from django.core.management.base import BaseCommand
 from django.db.models.query_utils import Q
 
-from tunga_tasks.notifications import remind_progress_event, notify_missed_progress_event
+from tunga_tasks.notifications.generic import remind_progress_event, notify_missed_progress_event
 from tunga_tasks.models import Task, ProgressEvent
 from tunga_tasks.tasks import initialize_task_progress_events
